+++ gulpfile.js
@@ -4,12 +4,14 @@ const { spawn } = require('child_process'),
       replace = require('gulp-replace'),
       clean = require('gulp-clean'),
       fs = require('fs'),
+      hash = require('gulp-hash-filename'),
       sass = require('gulp-dart-sass');
 
-gulp.task('frontend:rename', function(cb) {
+gulp.task('ui:patch', function(cb) {
     let chunkVendorsHash = '';
     let appHash = '';
-    fs.readdirSync('src/frontend/dist/js').forEach((filename) => {
+    let themeHash = '';
+    fs.readdirSync('src/digi_edit/static/app').forEach((filename) => {
         const chunkVendorsMatch = filename.match(/chunk-vendors\.([a-zA-Z0-9]+)?\.js/);
         if (chunkVendorsMatch) {
             chunkVendorsHash = chunkVendorsMatch[1];
@@ -19,8 +21,15 @@ gulp.task('frontend:rename', function(cb) {
             appHash = appMatch[1];
         }
     });
+    fs.readdirSync('src/digi_edit/static').forEach((filename) => {
+        const themeMatch = filename.match(/theme\.([a-zA-Z0-9]+)?\.css/);
+        if (themeMatch) {
+            themeHash = themeMatch[1];
+        }
+    });
     pump([
         gulp.src('src/digi_edit/templates/ui.jinja2'),
+        replace(/static\/theme(\.[a-zA-Z0-9]+)?\.css/, 'static/theme.' + themeHash + '.css'),
         replace(/chunk-vendors(\.[a-zA-Z0-9]+)?\.js/, 'chunk-vendors.' + chunkVendorsHash + '.js'),
         replace(/app(\.[a-zA-Z0-9]+)?\.js/, 'app.' + appHash + '.js'),
         gulp.dest('src/digi_edit/templates')
@@ -57,15 +66,27 @@ gulp.task('frontend:build:production', function(cb) {
     builder.on('exit', cb);
 });
 
-gulp.task('frontend', gulp.series('frontend:build:production', 'frontend:copy.clean', 'frontend:copy', 'frontend:rename'));
+gulp.task('frontend', gulp.series('frontend:build:production', 'frontend:copy.clean', 'frontend:copy'));
 
-gulp.task('theme', function(cb) {
+gulp.task('theme:build', function(cb) {
     pump([
         gulp.src('src/theme/theme.scss'),
         sass(),
+        hash({
+            format: '{name}.{hash}{ext}',
+        }),
         gulp.dest('src/digi_edit/static'),
     ], cb);
 });
+
+gulp.task('theme:clean', function(cb) {
+    pump([
+        gulp.src('src/digi_edit/static/theme.*.css'),
+        clean()
+    ], cb);
+})
+
+gulp.task('theme', gulp.series('theme:clean', 'theme:build'))
 
 gulp.task('css:sanitize', (cb) => {
     pump([
@@ -74,16 +95,16 @@ gulp.task('css:sanitize', (cb) => {
     ], cb);
 });
 
-gulp.task('default', gulp.parallel('theme', 'css:sanitize', 'frontend'));
+gulp.task('default', gulp.series(gulp.parallel('theme', 'css:sanitize', 'frontend'), 'ui:patch'));
 
 gulp.task('watch', gulp.parallel('frontend:build:development', 'css:sanitize', 'theme', function(cb) {
-    gulp.watch('src/theme/**/*.*', gulp.series('theme'));
+    gulp.watch('src/theme/**/*.*', gulp.series('theme', 'ui:patch'));
     gulp.watch('src/frontend/dist/js/*.*',
         {
             delay: 1000,
             events: ['add', 'change']
         },
-        gulp.series('frontend:copy'),
+        gulp.series('frontend:copy', 'ui:patch'),
     );
     cb();
 }));